View handling, API calls rework

+++ Server/1.3.b/api/request/add.php
@@ -17,6 +17,12 @@ include_once "../foreign/Vendor.php";
 //Geomath alghoritm
 include_once "../foreign/Geomath.php";
 
+//Read from database API call URL
+define("apiReadCall", "http://localhost/accesspointmap/server/1.3.b/api/request/read.php");
+
+//Control the existing of the record
+$recordExist = false;
+
 $database = new Database();
 $db = $database->getConnection();
 
@@ -25,7 +31,7 @@ $accessPoint = new AccessPoint($db);
 //Get the JSON data (POST)
 $data = json_decode(file_get_contents("php://input"));
 
-// verify the data
+//Verify the data
 if(
     !empty($data->bssid) &&
     !empty($data->ssid) &&
@@ -38,28 +44,72 @@ if(
     !empty($data->lowLongitude) &&
     !empty($data->security)
 ) {
-    $accessPoint->bssid = $data->bssid;
-    $accessPoint->ssid = $data->ssid;
-    $accessPoint->freq = $data->freq;
-    $accessPoint->signalLevel = $data->signalLevel;
-    $accessPoint->latitude = $data->latitude;
-    $accessPoint->longitude = $data->longitude;
-    $accessPoint->lowSignalLevel = $data->lowSignalLevel;
-    $accessPoint->lowLatitude = $data->lowLatitude;
-    $accessPoint->lowLongitude = $data->lowLongitude;
-    $accessPoint->signalArea = GeoMath::areaByGeoloc($data->latitude, $data->longitude, $data->lowLatitude, $data->lowLongitude);
-    $accessPoint->security = $data->security;
-    $accessPoint->vendor = Vendor::getVendor($data->bssid);
 
-    if($accessPoint->add())
-    {
-        //Object created and placed in the database
-        http_response_code(201);
-        echo json_encode(array("response" => "OK"));
-    } else {
-        //Service unavailable
-        http_response_code(503);
-        echo json_encode(array("response" => "Operation failed"));
+    //Calling the API to get all accessPoints to compare
+    $ch = curl_init();
+    curl_setopt($ch, CURLOPT_URL, apiReadCall);
+    curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+    curl_setopt($ch, CURLOPT_POST, true);
+    $response = json_decode(curl_exec($ch), true);
+    curl_close($ch);
+
+    //Check if record with this BSSID exist, decision whether to add or update
+    if($response["message"] != "No items found!") {
+        for($i = 0; $i < count($response["records"]); $i++) {
+            if($data->bssid == $response["records"][$i]["bssid"]) {
+                //RECORD UPDATE!  
+                $recordExist = true;
+
+                //Set the BSSID of the AccessPoint to be edited
+                $accessPoint->bssid = $data->bssid;
+
+                $accessPoint->signalLevel = $data->signalLevel;
+                $accessPoint->latitude = $data->latitude;
+                $accessPoint->longitude = $data->longitude;
+                $accessPoint->lowSignalLevel = $data->lowSignalLevel;
+                $accessPoint->lowLatitude = $data->lowLatitude;
+                $accessPoint->lowLongitude = $data->lowLongitude;
+                $accessPoint->signalArea = GeoMath::getArea($data->latitude, $data->longitude, $data->lowLatitude, $data->lowLongitude);
+
+                //Updata the AccessPoint
+                if($accessPoint->update()) {
+                    //Data sucessful edited
+                    http_response_code(200);
+                    echo json_encode(array("message" => "Object updated in database"));
+                } else {
+                    //Service unavailable
+                    http_response_code(501);
+                    echo json_encode(array("message" => "Operation failed!"));
+                }
+            }
+        }
+    }
+
+    if(!$recordExist) {
+        //RECORD CREATE
+
+        $accessPoint->bssid = $data->bssid;
+        $accessPoint->ssid = $data->ssid;
+        $accessPoint->freq = $data->freq;
+        $accessPoint->signalLevel = $data->signalLevel;
+        $accessPoint->latitude = $data->latitude;
+        $accessPoint->longitude = $data->longitude;
+        $accessPoint->lowSignalLevel = $data->lowSignalLevel;
+        $accessPoint->lowLatitude = $data->lowLatitude;
+        $accessPoint->lowLongitude = $data->lowLongitude;
+        $accessPoint->signalArea = GeoMath::getArea($data->latitude, $data->longitude, $data->lowLatitude, $data->lowLongitude);
+        $accessPoint->security = $data->security;
+        $accessPoint->vendor = Vendor::getVendor($data->bssid);
+
+        if($accessPoint->add()) {
+            //Object created and placed in the database
+            http_response_code(201);
+            echo json_encode(array("response" => "Object added to database"));
+        } else {
+            //Service unavailable
+            http_response_code(503);
+            echo json_encode(array("response" => "Operation failed"));
+        }
     }
 } else {
     //Bad data
